Move dashboard component to app specific dashboard.py
